requirements.txt
opencv-python
mediapipe
face_recognition
numpy
fastapi
uvicorn
sqlalchemy
python-dotenv
openai-whisper
pyaudio
requests
python-multipart
# AI integrations
openai
anthropic
# Optional but recommended
rich
tqdm
